V5 dual-loss ablation (null result) and v6 N=5k Gaussian bump.
v5: added an inverted-sRGB L1 term alongside the log-HDR L1 loss. No
visual lift over v4; reverted to the single log-HDR L1 loss and stripped
the dual-loss plumbing. v6: bumped the per-scene Gaussian cap from 3000
to 5000 (regen_data_n5k.sh + train_phase2_v6.sh). Also no measurable
quality change. Both null results recorded in PROGRESS.md sections 11
and 12; v4 baseline stands. Includes the .gitignore tidy for the v5/v6
checkpoint dirs and n5k data outputs.

+++ eval_val_set.py
@@ -62,7 +62,7 @@ def main():
             with torch.no_grad():
                 out = pipeline(
                     gaussians=gaussians, mask=mask, c2w=c2w, fov=fov,
-                    resolution=args.resolution, torch_dtype=torch.float32,
+                    resolution=args.resolution, torch_dtype=torch.bfloat16,
                 )
             hdr = out[0, 0].cpu().float().numpy()
             if tone_mapper is not None:
